BobmooAutoCrawlling/requirements.txt
requests
beautifulsoup4
lxml
charset-normalizer
google-generativeai
jsonschema
selenium
webdriver-manager
python-dotenv
